# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DrimPay API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformStats {
  totalTransactions: number;
  totalVolume: string;
  supportedCountries: number;
  activePartners: number;
  uptimePercent: number;
  merchantsOnboarded: number;
}

export interface BlogArticle {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  authorTitle: string;
  publishedAt: string;
  readingTimeMinutes: number;
  imageUrl?: string;
  tags: string[];
}

export interface BlogArticleList {
  articles: BlogArticle[];
  total: number;
  page: number;
  totalPages: number;
}

export interface BlogCategory {
  name: string;
  slug: string;
  count: number;
}

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  "full-time": "full-time",
  "part-time": "part-time",
  contract: "contract",
  internship: "internship",
} as const;

export interface Job {
  id: number;
  title: string;
  department: string;
  location: string;
  type: JobType;
  remote: boolean;
  description: string;
  requirements: string[];
  responsibilities: string[];
  postedAt: string;
}

export interface ContactFormBody {
  name: string;
  email: string;
  company?: string;
  subject: string;
  message: string;
}

export interface ContactFormResponse {
  success: boolean;
  message: string;
}

export type ServiceStatusStatus =
  (typeof ServiceStatusStatus)[keyof typeof ServiceStatusStatus];

export const ServiceStatusStatus = {
  operational: "operational",
  degraded: "degraded",
  outage: "outage",
  maintenance: "maintenance",
} as const;

export interface ServiceStatus {
  name: string;
  status: ServiceStatusStatus;
  uptimePercent: number;
  latencyMs: number;
}

export type ServiceStatusListOverall =
  (typeof ServiceStatusListOverall)[keyof typeof ServiceStatusListOverall];

export const ServiceStatusListOverall = {
  operational: "operational",
  degraded: "degraded",
  outage: "outage",
  maintenance: "maintenance",
} as const;

export interface ServiceStatusList {
  overall: ServiceStatusListOverall;
  services: ServiceStatus[];
  lastUpdated: string;
}

export type IncidentStatus =
  (typeof IncidentStatus)[keyof typeof IncidentStatus];

export const IncidentStatus = {
  investigating: "investigating",
  identified: "identified",
  monitoring: "monitoring",
  resolved: "resolved",
} as const;

export type IncidentSeverity =
  (typeof IncidentSeverity)[keyof typeof IncidentSeverity];

export const IncidentSeverity = {
  minor: "minor",
  major: "major",
  critical: "critical",
} as const;

export interface Incident {
  id: number;
  title: string;
  description: string;
  status: IncidentStatus;
  severity: IncidentSeverity;
  startedAt: string;
  resolvedAt?: string;
  affectedServices: string[];
}

export type PartnerType = (typeof PartnerType)[keyof typeof PartnerType];

export const PartnerType = {
  "mobile-money": "mobile-money",
  bank: "bank",
  fintech: "fintech",
  aggregator: "aggregator",
} as const;

export interface Partner {
  id: number;
  name: string;
  type: PartnerType;
  country: string;
  logoUrl?: string;
  description: string;
  website?: string;
}

export interface Operator {
  name: string;
  type: string;
  active: boolean;
}

export interface Country {
  code: string;
  name: string;
  flag: string;
  currency: string;
  operators: Operator[];
  payinEnabled: boolean;
  payoutEnabled: boolean;
}

export type ListBlogArticlesParams = {
  category?: string;
  page?: number;
  limit?: number;
};

export type ListJobsParams = {
  department?: string;
  location?: string;
};