# lib/api-zod/src/generated/types/contactFormBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DrimPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactFormBody {
  name: string;
  email: string;
  company?: string;
  subject: string;
  message: string;
}