# lib/api-zod/src/generated/types/country.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DrimPay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Operator } from "./operator";

export interface Country {
  code: string;
  name: string;
  flag: string;
  currency: string;
  operators: Operator[];
  payinEnabled: boolean;
  payoutEnabled: boolean;
}