# lib/api-zod/src/generated/types/operator.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DrimPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Operator {
  name: string;
  type: string;
  active: boolean;
}