# lib/api-zod/src/generated/types/partner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DrimPay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PartnerType } from "./partnerType";

export interface Partner {
  id: number;
  name: string;
  type: PartnerType;
  country: string;
  logoUrl?: string;
  description: string;
  website?: string;
}