# lib/api-zod/src/generated/types/serviceStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DrimPay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceStatusStatus } from "./serviceStatusStatus";

export interface ServiceStatus {
  name: string;
  status: ServiceStatusStatus;
  uptimePercent: number;
  latencyMs: number;
}